+++ lib/hatchet.rb
@@ -19,6 +19,7 @@ require 'hatchet/git_app'
 require 'hatchet/config'
 require 'hatchet/api_rate_limit'
 require 'hatchet/init_project'
+require 'hatchet/heroku_run'
 
 module Hatchet
   RETRIES = Integer(ENV['HATCHET_RETRIES']   || 1)